+++ src/ladb_opencutlist/ruby/helper/face_triangles_helper.rb
@@ -2,7 +2,6 @@ module Ladb::OpenCutList
 
   require_relative 'layer_visibility_helper'
 
-  require_relative '../utils/point3d_utils'
   require_relative '../utils/transformation_utils'
 
   module FaceTrianglesHelper
@@ -12,7 +11,7 @@ module Ladb::OpenCutList
     def _compute_children_faces_triangles(entities, transformation = nil, filtered_faces = nil)
       triangles = []
       entities.each { |entity|
-        next if entity.is_a?(Sketchup::Edge)   # Minor Speed improvement when there's a lot of edges
+        next if entity.is_a?(Sketchup::Edge)   # Minor Speed improvement when there are a lot of edges
         if entity.visible? && _layer_visible?(entity.layer)
           if entity.is_a?(Sketchup::Face)
             triangles.concat(_compute_face_triangles(entity, transformation)) if filtered_faces.nil? || filtered_faces.include?(entity)
@@ -38,16 +37,16 @@ module Ladb::OpenCutList
       mesh = face.mesh(0) # POLYGON_MESH_POINTS
       points = mesh.points
 
-      Point3dUtils::transform_points(points, transformation)
+      points.each { |point| point.transform!(transformation) } if transformation.is_a?(Geom::Transformation)
 
       triangles = []
-      mesh.polygons.each { |polygon|
-        polygon.each { |index|
+      mesh.polygons.each do |polygon|
+        polygon.each do |index|
           # Indicies start at 1 and can be negative to indicate edge smoothing.
           # Must take this into account when looking up the points in our array.
           triangles << points[index.abs - 1]
-        }
-      }
+        end
+      end
 
       triangles
     end