+++ src/ladb_opencutlist/ruby/helper/material_attributes_caching_helper.rb
@@ -7,9 +7,8 @@ module Ladb::OpenCutList
     def _get_material_attributes(material)
       material = Sketchup.active_model.materials[material.to_s] unless material.is_a?(Sketchup::Material)
       key = material ? material.name : '$EMPTY$'
-      @material_attributes_cache = {} unless @material_attributes_cache.is_a?(Hash)
-      @material_attributes_cache[key] = MaterialAttributes.new(material, true) unless @material_attributes_cache.has_key?(key)
-      @material_attributes_cache[key]
+      @material_attributes_cache ||= {}
+      @material_attributes_cache[key] ||= MaterialAttributes.new(material, true)
     end
 
   end