+++ src/ladb_opencutlist/ruby/utils/hash_utils.rb
@@ -5,9 +5,7 @@ module Ladb::OpenCutList
     def self.symbolize_keys(hash)
       return hash.transform_keys { |k| k.to_sym } if hash.respond_to?(:transform_keys)
       # Workaround for Ruby prior to 2.5
-      h = {}
-      hash.each_pair { |key, value| h[key.to_sym] = value }
-      h
+      hash.each_with_object({}) { |(key, value), h| h[key.to_sym] = value }
     end
 
   end